Crate amafatt

amafatt - processes the data from Amazon txt file and send an api json request to FattureInCloud.it

amafatt

Una applicazione prototipo per inserire dati ricevuti da Amazon in formato txt nella API di FattureInCloud.it.
Max sa cosa vuole dalla applicazione ed io (Luciano) costruirò la app come prototipo.
Dopo si vedrà se saremmo contenti con il risultato e si andrà avanti.

FattureInCloud.it

Ho creato un nuovo account per luciano.bestia@gmail.com dove si può sperimentare.
Io e Max abbiamo accesso a questo account e possiamo collaborare.
La documentazione della API e qui:
https://api.FattureInCloud.it/v1/documentation/dist/#!/Richiesta_generica/jsonrequest
La lingua di tutti i programmatori e l'Inglese e per questo ci sarà tanto inglese nella source code.
La source code e disponibile sul mio GitHub privato https://github.com/LucianoBestia/amafatt solo per i collaboratori.

Rust, Wasm, Dodrio

The programming language of this application is not important.
It is easy to rewrite the application in any other language if necessary.
The important thing is the application logic and mapping of data.
I will create the prototype in the Rust programming language. It will be compiled in Wasm/WebAssembly that runs inside all modern browsers.
The GUI interface will be in simple HTML updated with Dodrio Virtual Dom.
There will not be necessary any special servers, JavaScript, npm, web-pack or complicated css and html.

Development environment

Install the rust toolchain from
https://www.rust-lang.org/tools/install.
For wasm workflow tool installation on Windows:
https://rustwasm.github.io/wasm-pack/installer/
You will need also a simple static file server:
cargo install basic-http-server

Source

Absolutely everything manually coded is in Rust language in the folder /src/.
No manual JavaScript or HTML needed.
The index.html file is a simple template.
The pkg/mem.js file is generated by wasm-bindgen on building.
I prepared a small simple css with css grid for styling based on w3.css.
No need for npm or web-pack.
The proper way to do json would be with structs and vectors, but it is so much typing and retyping of datatypes and input fields - for small benefit. I will use whole json as a string. And there replace what should be replaced. Much simpler that way for this project. Even json objects are too complicated. I will use all that as simple strings.

Make

I prepared all the needed tasks/flows in cargo make.
https://github.com/sagiegurari/cargo-make
Install it simply with cargo install --force cargo-make
cargo make - lists the possible available/public flows/tasks
cargo make dev - builds the development version and runs the server and the browser
cargo make release - builds the release version and runs the server and the browser
The make script also lunches the basic-http-serverand open the browser.
TODO: could make a flow to publish to GitHub or to google vm or a task for increment version numbers.

Workflow

There is only one screen with big text fields for request and response from FattureInCloud.it.
A menu will have buttons to create different requests.
For importing data from Amazon there will be another text field where to copy/paste the content of the txt file from Amazon. No file uploads here because we don't really need/have a server.
The button Extract order-id will create a vector of order-id in case there are multiples in the amazon txt. For every order-id there will be a new button to process the txt and create a json string as close as possible what the API needs. That json data is possible to manually edit if needed.
The button Send json sends the json request to the API and fetch the response into the response field. Only one order-id should be sent at a time.
After that all needed editing, deleting, changing is made through the FattureInCloud.it user interface.
TODO: after some time it can be modified to send all the order-id one by one automatically.

reverse proxy

Modern browsers don't allow cross origin fetch by default.
The server must allow it explicitly with a special header Access-Control-Allow-Origin=*.
As expected, the site FattureInCloud.it does not send this header.
I wrote them to correct this, but I don't have any hope they will listen to me.
The workaround is a reverse proxy that receives the original request, sends it to FattureInCloud.it, receives the response, add this special header and sends the response back to the browser. I use nginx on my google cloud vm. To achieve this I added this lines in the configuration file \etc\nginx\sites-available\default:

	#region amafatt
 
		location /v1/ {
			proxy_pass https://api.FattureInCloud.it/v1/;
			proxy_buffering off;
			add_header  Access-Control-Allow-Origin *;
		}
	#endregion

All the requests will be now sent to http://34.87.17.103/v1/.

Reference

https://github.com/rustwasm/wasm-bindgen/tree/master/examples/fetch
https://api.FattureInCloud.it/v1/documentation/dist/#!/Richiesta_generica/jsonrequest

Modules

amazonmod
amazonmod - generic HashMap to deserialize the amazon csv

credentialsmod
credentialsmod.rs - the credentials are saved in the LocalStorage of the browser

fetchmod
fetchmod - isolate/encapsulate fetch api in a module because it is all async

isocountriesmod
isocountriesmod - country names and iso code

reqbodymod
reqbodymod - all request body are created in this module

rootrenderingmod
rootrenderingmod - Root Rendering Component is the base struct for Dodrio Virtual Dom
rrc contains all the data needed for UI and reacts to all the events of the UI.
Other subcomponents can be more or less smart, with or without their own data, but that data are only copied values from the rrc component.

Functions

log1
simple console write with a string

wasm_bindgen_start
this is the start function that wasm_bindgen calls